+++ src/favorites/favorites.module.ts
@@ -4,9 +4,9 @@ https://docs.nestjs.com/modules
 
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { FavoritesController } from './favorites.controller';
+import { FavoritesController } from './controller/favorites.controller';
 import { FavoriteSchema } from './schemas/favorites.schema';
-import { FavoriteService } from './shared/favorite.service';
+import { FavoriteService } from './service/favorite.service';
 
 @Module({
   imports: [
